fix: 알림 페이지에 useNoticesQuery 적용 및 폴더명 수정
